+++ Examples/Objective-C/Navigine Example/ViewController.h
@@ -14,7 +14,8 @@
 <
 UIScrollViewDelegate,
 NavigineCoreDelegate,
-NavigineCoreNavigationDelegate
+NavigineCoreNavigationDelegate,
+NavigineCoreLocationDelegate
 >
 @property (weak, nonatomic) IBOutlet UIImageView *imageView;
 @property (weak, nonatomic) IBOutlet UIScrollView *scrollView;